+++ integration-test/src/main.rs
@@ -1,7 +1,7 @@
 mod const_definition;
 mod help;
 mod tests_omni_lock;
-mod tests_otx_lock;
+mod tests_otx_sighash_lock;
 mod utils;
 
 use help::{setup, teardown};
